+++ addOns/ascanrulesAlpha/CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
  - Terminology
 
+### Removed
+- The following scan rules were removed and promoted to Beta: Cloud Meta Data, .env File, Hidden Files, XSLT Injection (Issue 6211).
+
 ## [29] - 2020-08-13
 ### Changed
 - Maintenance changes.